+++ kernel/core/log.c
@@ -1,15 +1,15 @@
-#include <mvos/serial.h>
+#include <mvos/console.h>
 
 void klog(const char *message) {
-    serial_write_string(message);
+    console_write_string(message);
 }
 
 void klogln(const char *message) {
-    serial_write_string(message);
-    serial_write_string("\n");
+    console_write_string(message);
+    console_write_string("\n");
 }
 
 void klog_u64(uint64_t value) {
-    serial_write_u64(value);
-    serial_write_string("\n");
+    console_write_u64(value);
+    console_write_string("\n");
 }